Curso Spring Boot - 08 de Jan de 2020

+++ src/main/java/com/nelioalves/cursomc/domain/Categoria.java
@@ -10,6 +10,8 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.ManyToMany;
 
+import com.fasterxml.jackson.annotation.JsonManagedReference;
+
 @Entity
 public class Categoria implements Serializable {
 	private static final long serialVersionUID = 1L;
@@ -17,7 +19,7 @@ public class Categoria implements Serializable {
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	private Integer id;
 	private String nome;
-		
+	
 	@ManyToMany(mappedBy = "categorias")
 	private List<Produto> produtos = new ArrayList<>();
 		
@@ -30,7 +32,7 @@ public class Categoria implements Serializable {
 		this.nome = nome;
 	}
 
-	public long getId() {
+	public Integer getId() {
 		return id;
 	}
